trying to fix java bindings compile errors
