--- modules/3d/misc/objc/gen_dict.json
@@ -1,2 +1,0 @@
-{
-}
