+++ modules/java/generator/templates/java_module.prolog
@@ -1,7 +1,7 @@
 //
 // This file is auto-generated. Please don't modify it!
 //
-package org.opencv.$module;
+package org.opencv.$jmodule;
 
 $imports
